+++ changelog.md
@@ -1,5 +1,10 @@
-# 1.0.0
-- Initial release
+# 1.0.4
+- Improved button coloring
+- Freeze button
 
 # 1.0.1
 - Mac OS support
+
+# 1.0.0
+- Initial release
+